+++ Sources/Extensions/Notification.swift
@@ -1,5 +1,5 @@
 //
-//  Notification+Rover.swift
+//  Notification.swift
 //  Rover
 //
 //  Created by Andrew Clunis on 2019-03-18.
