+++ Sources/Models/BarcodeBlock.swift
