+++ Sources/UI/Views/BarcodeCell.swift
